requirements.txt
gradio>=4.44.0
openai>=1.30.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
faster-whisper>=1.0.0
f5-tts==0.1.1
numpy>=1.24.0